UPSTREAM: <carry>: Drop validation-gen
In https://github.com/kubernetes/kubernetes/pull/130349 for KEP-5073
validation-gen was added to `gen_helpers`.

KCP is currently not using validation-gen, but it is causing errors as
validation-gen directly imports k8s.io/apimachinery, which is rewritten
in the go.mod to `../apimachinery`.

This is working fine locally and in go workspaces, but breaks as soon as
`k8s.io/code-generator` is used from the kcp-dev/kubernetes fork, as
then the rewrite still exists but points to nothing.

Upstream the rewrites are resolved to the respective commit in the
module when publishing the changes to their repositories.

Signed-off-by: Nelo-T. Wallus <n.wallus@sap.com>